Fix stats to new buffer
